Fix digigr8 camlib compilation warnings.
Resolves number of signed-unsigned comparison issues.

+++ camlibs/digigr8/digi_postprocess.c
@@ -336,7 +336,7 @@ static int
 histogram (unsigned char *data, unsigned int size, int *htable_r,
 					    int *htable_g, int *htable_b)
 {
-	int x;
+	unsigned int x;
 	/* Initializations */
 	for (x = 0; x < 0x100; x++) {
 		htable_r[x] = 0;
@@ -424,7 +424,7 @@ white_balance (unsigned char *data, unsigned int size, float saturation)
 			r_factor=%1.3f, g_factor=%1.3f, b_factor=%1.3f\n",
 					r, g, b, r_factor, g_factor, b_factor);
 	if (max_factor <= 1.4) {
-		for (x = 0; x < (size * 3); x += 3)
+		for (x = 0; x < ((int)size * 3); x += 3)
 		{
 			d = (data[x + 0] << 8) * r_factor + 8;
 			d >>= 8;
@@ -462,7 +462,7 @@ white_balance (unsigned char *data, unsigned int size, float saturation)
 			r_factor=%1.3f, g_factor=%1.3f, b_factor=%1.3f\n",
 				r, g, b, r_factor, g_factor, b_factor);
 
-	for (x = 0; x < (size * 3); x += 3)
+	for (x = 0; x < ((int)size * 3); x += 3)
 	{
 		d = (int) 0xff08 - (((0xff - data[x + 0]) << 8) * r_factor);
 		d >>= 8;
@@ -484,7 +484,7 @@ white_balance (unsigned char *data, unsigned int size, float saturation)
 	/* ------------------ COLOR ENHANCE ------------------ */
 
 	if(saturation > 0.0) {
-		for (x = 0; x < (size * 3); x += 3)
+		for (x = 0; x < ((int)size * 3); x += 3)
 		{
 			r = data[x+0]; g = data[x+1]; b = data[x+2];
 			d = (int) (r + g + b) / 3.;
